ART: Fix old warnings
Fix Wconstant-conversion warnings.

Partially reverts commit df53be273509dd43725870fb20a2c7d71f7fbfd3.

Bug: 28149048
Bug: 29823425
Test: m
Test: m test-art-host

+++ compiler/optimizing/code_generator_arm_vixl.cc
@@ -2662,7 +2662,10 @@ void CodeGeneratorARMVIXL::GenerateFrameEntry() {
     vixl32::Register temp = temps.Acquire();
     // Initialize should_deoptimize flag to 0.
     __ Mov(temp, 0);
-    GetAssembler()->StoreToOffset(kStoreWord, temp, sp, -kShouldDeoptimizeFlagSize);
+    GetAssembler()->StoreToOffset(kStoreWord,
+                                  temp,
+                                  sp,
+                                  -static_cast<int32_t>(kShouldDeoptimizeFlagSize));
   }
 
   int adjust = GetFrameSize() - FrameEntrySpillSize();
